Conflicts:
	Pyssypeli/Assembly-CSharp-Editor-vs.csproj
	Pyssypeli/Assembly-CSharp-Editor.csproj
	Pyssypeli/Assembly-CSharp-firstpass-vs.csproj
	Pyssypeli/Assembly-CSharp-firstpass.csproj
	Pyssypeli/Assembly-CSharp-vs.csproj
	Pyssypeli/Assembly-CSharp.csproj
	Pyssypeli/Pyssypeli-csharp.sln
	Pyssypeli/Pyssypeli.sln

+++ jotain/human.cs
@@ -11,8 +11,8 @@ public class human : MonoBehaviour
 		private int hp ;
 		private int energy ;
 		private int money ;
-		private float locationX ;
-		private float locationY ;
+		private int locationX ;
+		private int locationY ;
 		private int agility ;
 		private int constitution ;
 		private int luck ;
@@ -31,7 +31,7 @@ public class human : MonoBehaviour
 	// Update is called once per frame
 	void Update () 
 	{
-	  
+	
 	}
 	
 	public void moveUp()
